getAssets方法添加资源注入，与解决baseapk无法访问插件资源问题
